+++ plugin/src/main/groovy/org/openbakery/KeychainCreateTask.groovy
@@ -30,13 +30,10 @@ class KeychainCreateTask extends AbstractXcodeTask {
 	@TaskAction
 	def create() {
 
+
+
 		if (project.xcodebuild.sdk.startsWith("iphonesimulator")) {
 			println("The simulator build does not need a provisioning profile")
-			return
-		}
-
-		if (project.xcodebuild.signing.certificateURI == null) {
-			println("not certificateURI specifed so do not create the keychain")
 			return
 		}
 
@@ -47,8 +44,11 @@ class KeychainCreateTask extends AbstractXcodeTask {
 		}
 
 		if (project.xcodebuild.signing.certificateURI == null) {
-			throw new InvalidUserDataException("Property project.xcodebuild.signing.certificateURI is missing")
+			println("not certificateURI specifed so do not create the keychain")
+			return
 		}
+
+
 		if (project.xcodebuild.signing.certificatePassword == null) {
 			throw new InvalidUserDataException("Property project.xcodebuild.signing.certificatePassword is missing")
 		}
@@ -62,12 +62,27 @@ class KeychainCreateTask extends AbstractXcodeTask {
 		if (!new File(keychainPath).exists()) {
 			runCommand(["security", "create-keychain", "-p", project.xcodebuild.signing.keychainPassword, keychainPath])
 		}
-
-		//runCommand(["security", "default-keychain", "-s", getKeychainName()])
 		runCommand(["security", "-v", "import", certificateFile, "-k", keychainPath, "-P", project.xcodebuild.signing.certificatePassword, "-T", "/usr/bin/codesign"])
 
 
-		//runCommand(["security", "list"])
+		if (getOSVersion().minor >= 9) {
+			String keychainList = runCommandWithResult(["security", "list-keychains"]);
+
+			def commandList = [
+							"security",
+							"list-keychains",
+							"-s"
+			]
+			for (String keychain in keychainList.split("\n")) {
+				String trimmedKeychain = keychain.replaceAll(/^\s*\"|\"$/, "");
+				if (new File(trimmedKeychain).exists()) {
+					commandList.add(trimmedKeychain);
+				}
+
+			}
+			commandList.add(keychainPath)
+			runCommand(commandList)
+		}
 	}